new home for comment strategies

+++ src/Parser/DomainLiteral.php
@@ -146,11 +146,11 @@ class DomainLiteral extends Parser
 
         // Extract IPv4 part from the end of the address-literal (if there is one)
         if ($IPv4Match > 0) {
-            $index = strrpos($addressLiteralIPv4, $matchesIP[0]);
+            $index = (int) strrpos($addressLiteralIPv4, $matchesIP[0]);
             //There's a match but it is at the start
             if ($index > 0) {
                 // Convert IPv4 part to IPv6 format for further testing
-                return substr($addressLiteralIPv4, 0, (int) $index) . '0:0';
+                return substr($addressLiteralIPv4, 0, $index) . '0:0';
             }
         }
